+++ Core2Identity/Controllers/UlkeController.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Core2Identity.Models;
 using GeziYazisiSitesi.Modals;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Core2Identity.Controllers
@@ -18,20 +19,35 @@ namespace Core2Identity.Controllers
         }
         public IActionResult Index()
         {
-            return View(context.Ulke);
+            if (HttpContext.Session.GetInt32("Yetki") == 1)
+            {
+                return View(context.Ulke);
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
+                
         }
 
         [HttpGet]
         public IActionResult Create()
         {
-            return View();
+            if (HttpContext.Session.GetInt32("Yetki") == 1)
+            {
+                return View();
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
 
         [HttpPost]
         [ValidateAntiForgeryToken]
         public IActionResult Create(Ulke entity)
         {
-            if (ModelState.IsValid)
+            if (ModelState.IsValid && HttpContext.Session.GetInt32("Yetki") == 1)
             {
                 context.Ulke.Add(entity);
                 context.SaveChanges();
@@ -56,7 +72,7 @@ namespace Core2Identity.Controllers
         [ValidateAntiForgeryToken]
         public IActionResult Edit(Ulke entity)
         {
-            if (ModelState.IsValid)
+            if (ModelState.IsValid && HttpContext.Session.GetInt32("Yetki") == 1)
             {
 
                 var ulke = context.Ulke.FirstOrDefault(i => i.UlkeId == entity.UlkeId);
@@ -83,7 +99,7 @@ namespace Core2Identity.Controllers
         public IActionResult DeleteConfirmed(int id)
         {
             var ulke = context.Ulke.FirstOrDefault(i => i.UlkeId == id);
-            if (ulke != null)
+            if (ulke != null && HttpContext.Session.GetInt32("Yetki") == 1)
             {
                 context.Ulke.Remove(ulke);
                 context.SaveChanges();
